requirements.txt
requests==2.28.2
confluent-kafka==2.0.2
uuid7==0.1.0
isort==5.12.0
black==23.9.1
flake8==6.1.0
pytest==7.4.2
python-dotenv==1.0.0
openmetadata-ingestion==1.4.5.0